refactor: Chart 및 Index 관련 Controller, Service 통합 및 코드 정리
- ChartController, IndexRankController 삭제 및 관련 메서드 DashboardController로 통합
- ChartService, IndexRankService 내용 DashboardService로 이전 및 정리
- 불필요한 클래스 및 메서드 삭제, 코드 가독성 향상
